fix(permission): fix permission id data type

+++ src/main/java/com/studies/algafood/domain/repository/PermissionRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface PermissionRepository {
 
     List<Permission> list();
-    Permission find(Integer id);
+    Permission find(Long id);
     Permission save(Permission permission);
     void remove(Permission permission);
 }
